requirements.txt
groq>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-search-results>=2.4.2
PyGithub>=2.1.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
markdown>=3.5.0
streamlit>=1.32.0
python-dotenv>=1.0.0
lxml>=5.1.0
